+++ blocklib/blocks/test/cuda/qa_multiply_const.py
@@ -7,7 +7,7 @@
 #
 
 
-from newsched import gr, gr_unittest, blocks, math
+from gnuradio import gr, gr_unittest, blocks, math
 # import pmt
