+++ blocklib/fft/test/qa_fft.py
@@ -9,7 +9,7 @@
 #
 
 
-from newsched import gr, gr_unittest, fft, blocks
+from gnuradio import gr, gr_unittest, fft, blocks
 
 # Note: Octave code to verify these results:
 # primes = [2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 127, 131, 137, 139, 149, 151, 157, 163, 167,
