+++ blocklib/filter/python/filter/meson.build
@@ -21,7 +21,7 @@ endforeach
 d = {
   'blocks' : filter_pybind_names,
   'module' : 'filter',
-  'imports' : ['newsched.gr','newsched.fft']
+  'imports' : ['gnuradio.gr','newsched.fft']
 }
 
 gen_filter_pybind = custom_target('gen_filter_pybind',
@@ -42,7 +42,7 @@ newsched_blocklib_filter_pybind = py3_inst.extension_module('filter_python',
     dependencies : [newsched_blocklib_filter_dep, python3_dep, pybind11_dep],
     link_language : 'cpp',
     install : true,
-    install_dir : join_paths(py3_inst.get_install_dir(),'newsched','filter')
+    install_dir : join_paths(py3_inst.get_install_dir(),'gnuradio','filter')
 )
 
 newsched_blocklib_filter_pybind_dep = declare_dependency(include_directories : incdir,
@@ -51,4 +51,4 @@ newsched_blocklib_filter_pybind_dep = declare_dependency(include_directories : i
 
 
 # Target for pure python
-py3_inst.install_sources(srcs, subdir : join_paths('newsched','filter'))
+py3_inst.install_sources(srcs, subdir : join_paths('gnuradio','filter'))
