+++ blocklib/math/add/numpy/add.py
@@ -1,5 +1,5 @@
-from newsched import math
-from newsched import gr
+from gnuradio import math
+from gnuradio import gr
 
 class add_ff(math.add_ff):
     def __init__(self, *args, **kwargs):
