+++ blocklib/math/python/math/meson.build
@@ -22,7 +22,7 @@ endforeach
 d = {
   'blocks' : math_pybind_names,
   'module' : 'math',
-  'imports' : ['newsched.gr']
+  'imports' : ['gnuradio.gr']
 }
 
 gen_math_pybind = custom_target('gen_math_pybind',
@@ -43,7 +43,7 @@ newsched_blocklib_math_pybind = py3_inst.extension_module('math_python',
     dependencies : [newsched_blocklib_math_dep, python3_dep, pybind11_dep],
     link_language : 'cpp',
     install : true,
-    install_dir : join_paths(py3_inst.get_install_dir(),'newsched','math')
+    install_dir : join_paths(py3_inst.get_install_dir(),'gnuradio','math')
 )
 
 newsched_blocklib_math_pybind_dep = declare_dependency(include_directories : incdir,
@@ -52,7 +52,7 @@ newsched_blocklib_math_pybind_dep = declare_dependency(include_directories : inc
 
 
 # Target for pure python
-py3_inst.install_sources(srcs, subdir : join_paths('newsched','math'))
+py3_inst.install_sources(srcs, subdir : join_paths('gnuradio','math'))
 
 subdir('numpy')
-py3_inst.install_sources(files('numpy/__init__.py'), subdir : join_paths('newsched','math','numpy'))
+py3_inst.install_sources(files('numpy/__init__.py'), subdir : join_paths('gnuradio','math','numpy'))
