+++ blocklib/math/test/qa_conjugate.py
@@ -9,7 +9,7 @@
 #
 
 
-from newsched import gr, gr_unittest, blocks, math
+from gnuradio import gr, gr_unittest, blocks, math
 
 
 class test_conjugate (gr_unittest.TestCase):
