+++ blocklib/math/test/qa_type_conversions.py
@@ -9,7 +9,7 @@
 #
 
 
-from newsched import gr, gr_unittest, blocks, math
+from gnuradio import gr, gr_unittest, blocks, math
 
 from math import sqrt, atan2
