+++ grc/blocks/options.block.yml
@@ -133,7 +133,7 @@ asserts:
 
 templates:
     imports: |-
-        from newsched import gr
+        from gnuradio import gr
         #from gnuradio.filter import firdes
         #from gnuradio.fft import window
         import sys
