+++ grc/core/utils/epy_block_io.py
@@ -45,7 +45,7 @@ def _find_block_class(source_code, cls):
 
 def extract(cls):
     try:
-        from newsched import gr
+        from gnuradio import gr
         import pmt
     except ImportError:
         raise EnvironmentError("Can't import GNU Radio")
@@ -92,7 +92,7 @@ def extract(cls):
 if __name__ == '__main__':
     blk_code = """
 import numpy as np
-from newsched import gr
+from gnuradio import gr
 import pmt
 
 class blk(gr.sync_block):
