+++ grc/meson.build
@@ -6,19 +6,19 @@ subdir('scripts')
 
 install_subdir(
   'core',
-  install_dir         : join_paths(py3_inst.get_install_dir(),'newsched','grc'),
+  install_dir         : join_paths(py3_inst.get_install_dir(),'gnuradio','grc'),
   strip_directory     : false,
 )
 
 install_subdir(
   'converter',
-  install_dir         : join_paths(py3_inst.get_install_dir(),'newsched','grc'),
+  install_dir         : join_paths(py3_inst.get_install_dir(),'gnuradio','grc'),
   strip_directory     : false,
 )
 
 install_subdir(
   'gui',
-  install_dir         : join_paths(py3_inst.get_install_dir(),'newsched','grc'),
+  install_dir         : join_paths(py3_inst.get_install_dir(),'gnuradio','grc'),
   strip_directory     : false,
 )
 
@@ -29,4 +29,4 @@ py3_inst.install_sources([
     'compiler.py',
     'main.py',
 ], 
-subdir : join_paths('newsched','grc'))
+subdir : join_paths('gnuradio','grc'))
