+++ grc/scripts/gnuradio-companion
@@ -71,7 +71,7 @@ def check_gtk():
 
 def check_gnuradio_import():
     try:
-        from newsched import gr
+        from gnuradio import gr
     except ImportError as err:
         die(err, GR_IMPORT_ERROR_MESSAGE)
 
@@ -88,7 +88,7 @@ def run_main():
 
     if not script_path.endswith(source_tree_subpath):
         # run the installed version
-        from newsched.grc.main import main
+        from gnuradio.grc.main import main
     else:
         print("Running from source tree")
         sys.path.insert(1, script_path[:-len(source_tree_subpath)])
