+++ python/gnuradio/meson.build
@@ -7,4 +7,4 @@ configure_file(copy: true,
 )
 endforeach
 
-py3_inst.install_sources(files(srcs), subdir : join_paths('newsched'))
+py3_inst.install_sources(files(srcs), subdir : join_paths('gnuradio'))
