+++ schedulers/nbt/python/schedulers/nbt/bindings/meson.build
@@ -7,7 +7,7 @@ newsched_scheduler_nbt_pybind = py3_inst.extension_module('scheduler_nbt_python'
     dependencies : [newsched_gr_dep, newsched_scheduler_nbt_dep, python3_dep, pybind11_dep],
     link_language : 'cpp',
     install : true,
-    install_dir : join_paths(py3_inst.get_install_dir(),'newsched','schedulers','nbt')
+    install_dir : join_paths(py3_inst.get_install_dir(),'gnuradio','schedulers','nbt')
 )
 
 newsched_runtime_pybind_dep = declare_dependency(
