+++ schedulers/nbt/python/schedulers/nbt/meson.build
@@ -8,4 +8,4 @@ configure_file(copy: true,
 )
 endforeach
 
-py3_inst.install_sources(files('__init__.py'), subdir : join_paths('newsched','schedulers','nbt'))
+py3_inst.install_sources(files('__init__.py'), subdir : join_paths('gnuradio','schedulers','nbt'))
